pyproject.toml
[tool.poetry]
name = "billing"
version = "0.1.0"
description = "Billing programm"
authors = ["dyens <dyens@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.7.5"
alembic = "^1.3.0"
dynaconf = "^2.2.0"
aiohttp = "^3.6.2"
asyncpg = "^0.19.0"
psycopg2 = "^2.8.4"
asyncpgsa = "^0.26.1"

[tool.poetry.dev-dependencies]
mypy = "^0.720.0"
pytest = "^5.0"
sqlalchemy-stubs = "^0.2.0"
pytest-cov = "^2.7"
wemake-python-styleguide = "^0.12.5"
pytest-asyncio = "^0.10.0"
pytest-aiohttp = "^0.3.0"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
